HOLO-134 <refactor: children 컴포넌트 분리>
