+++ src/app/layout.tsx
@@ -1,5 +1,5 @@
 import "@/app/globals.css";
-import Header from "@/components/layout/Header";
+import MainContainer from "@/components/layout/MainContainer";
 import TQProvider from "@/components/providers/TQProvider";
 import type { Metadata } from "next";
 import "swiper/css";
@@ -20,11 +20,8 @@ const RootLayout = ({
   return (
     <html lang="ko" suppressHydrationWarning={true}>
       <body>
-        <Header />
         <TQProvider>
-          <main className="mb-[60px] h-[calc(100vh-120px)] overflow-auto lg:mb-0 lg:ml-[240px] lg:h-auto lg:overflow-visible">
-            {children}
-          </main>
+          <MainContainer>{children}</MainContainer>
         </TQProvider>
       </body>
     </html>
